fix(docker): ESM-Postbuild auch für session-export-report
tsc -b emittiert beide ESM-Workspaces; ohne Patch scheitert der
nächste Relative-Import in session-export-report zur Laufzeit.
